feat: refactor yml editor
