+++ packages/frontend/src/components/visual-editor/v4/components/yaml-editor/index.tsx
@@ -1,0 +1,8 @@
+/*
+ * Hexabot — Fair Core License (FCL-1.0-ALv2)
+ * Copyright (c) 2025 Hexastack.
+ * Full terms: see LICENSE.md.
+ */
+
+export { YamlEditor } from "./YamlEditor";
+export type { YamlEditorProps } from "./types";
