+++ packages/frontend/src/components/visual-editor/v4/components/yaml-editor/styles/yaml-editor.css
